Rename the file CountMazePaths.java to CountUniqueMazePaths.java in the source and test folders.
